# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Keystone
  description: Keystone is a double-entry general-ledger HTTP API. Money is integer
    minor units (ISO 4217); journal entries are balanced postings in a configured
    base currency; periods are closed sequentially. See the project README and ADRs
    for design rationale.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /periods/{yyyymm}/reopen:
    post:
      tags:
      - period-controller
      summary: Reopen the most recently closed period
      description: Reopens the given month. Only the most recently closed period can
        be reopened; reopening earlier ones is rejected to preserve the sequential
        close invariant.
      operationId: reopen
      parameters:
      - name: yyyymm
        in: path
        required: true
        schema:
          type: string
          pattern: "^\\d{4}-(0[1-9]|1[0-2])$"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /periods/{yyyymm}/close:
    post:
      tags:
      - period-controller
      summary: Close a period
      description: Closes the given month. Must close from the earliest open month
        with postings; closing out of order is rejected. Closed periods reject new
        postings.
      operationId: close
      parameters:
      - name: yyyymm
        in: path
        required: true
        schema:
          type: string
          pattern: "^\\d{4}-(0[1-9]|1[0-2])$"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /journal-entries:
    post:
      tags:
      - journal-entry-controller
      summary: Post a journal entry
      description: "Persists a balanced double-entry journal entry. Each posting carries\
        \ a transaction amount and a base-currency equivalent (baseMinorUnits); the\
        \ entry must balance in the configured base currency. Postings against an\
        \ unknown, inactive, or non-leaf account are rejected; so are postings into\
        \ a closed period."
      operationId: post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostJournalEntryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /accounts:
    get:
      tags:
      - account-controller
      summary: List all accounts
      description: "Returns every account in the chart, ordered by code."
      operationId: list
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountResponse"
    post:
      tags:
      - account-controller
      summary: Create an account
      description: Creates a new account in the chart of accounts. The code must be
        unique. If parentCode is supplied it must reference an existing account; the
        new account becomes a leaf under that parent.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /accounts/{code}/reactivate:
    post:
      tags:
      - account-controller
      summary: Reactivate an account
      description: Re-enables an account previously deactivated. Idempotent.
      operationId: reactivate
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /accounts/{code}/deactivate:
    post:
      tags:
      - account-controller
      summary: Deactivate an account
      description: Marks the account inactive. Inactive accounts reject new postings
        but historical entries referencing them remain unchanged. Idempotent.
      operationId: deactivate
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /accounts/{code}:
    get:
      tags:
      - account-controller
      summary: Fetch one account by code
      description: "Returns the account with the given code, or 404 if no such account\
        \ exists."
      operationId: get
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    patch:
      tags:
      - account-controller
      summary: Rename or re-parent an account
      description: Applies optional rename (newCode) and/or re-parent (newParentCode).
        Pass an empty newParentCode to detach the account from its current parent.
        Cycles are rejected.
      operationId: update
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /reports/trial-balance:
    get:
      tags:
      - trial-balance-controller
      summary: Trial balance
      description: "Read-only projection of postings, one row per (accountCode, currency)\
        \ pair with debits, credits, balance in both transaction and base currency.\
        \ asOf defaults to today (UTC); includeZero defaults to false (suppress net-zero\
        \ rows). Sorted by (accountCode, currency)."
      operationId: get_1
      parameters:
      - name: asOf
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: includeZero
        in: query
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrialBalanceRowResponse"
  /periods:
    get:
      tags:
      - period-controller
      summary: List periods
      description: "Without query params, returns an empty list (open periods are\
        \ implicit and not enumerable). With ?status=closed, returns every closed\
        \ period in chronological order."
      operationId: list_1
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PeriodResponse"
  /periods/{yyyymm}:
    get:
      tags:
      - period-controller
      summary: Fetch a period's status
      description: Returns the period's state for the given month (yyyy-MM). Months
        that have never been closed return status OPEN with no recorded activity.
      operationId: get_2
      parameters:
      - name: yyyymm
        in: path
        required: true
        schema:
          type: string
          pattern: "^\\d{4}-(0[1-9]|1[0-2])$"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
components:
  schemas:
    PostJournalEntryRequest:
      type: object
      properties:
        occurredOn:
          type: string
          format: date
        description:
          type: string
          maxLength: 500
          minLength: 0
        postings:
          type: array
          items:
            $ref: "#/components/schemas/PostingRequest"
          minItems: 1
      required:
      - description
      - occurredOn
      - postings
    PostingRequest:
      type: object
      properties:
        account:
          type: string
          minLength: 1
        side:
          type: string
          pattern: ^(DEBIT|CREDIT)$
        minorUnits:
          type: integer
          format: int64
        currency:
          type: string
          minLength: 1
          pattern: "^[A-Z]{3}$"
        baseMinorUnits:
          type: integer
          format: int64
      required:
      - account
      - currency
      - side
    CreateAccountRequest:
      type: object
      properties:
        code:
          type: string
          maxLength: 64
          minLength: 0
        name:
          type: string
          maxLength: 200
          minLength: 0
        type:
          type: string
          minLength: 1
          pattern: ^(ASSET|LIABILITY|EQUITY|REVENUE|EXPENSE)$
        currency:
          type: string
          minLength: 1
          pattern: "^[A-Z]{3}$"
        parentCode:
          type: string
          maxLength: 64
          minLength: 0
      required:
      - code
      - currency
      - name
      - type
    UpdateAccountRequest:
      type: object
      properties:
        newCode:
          type: string
          maxLength: 64
          minLength: 0
        newParentCode:
          type: string
          maxLength: 64
          minLength: 0
    TrialBalanceRowResponse:
      type: object
      properties:
        accountCode:
          type: string
        currency:
          type: string
        debits:
          type: integer
          format: int64
        credits:
          type: integer
          format: int64
        balance:
          type: integer
          format: int64
        baseDebits:
          type: integer
          format: int64
        baseCredits:
          type: integer
          format: int64
        baseBalance:
          type: integer
          format: int64
    PeriodResponse:
      type: object
      properties:
        yearMonth:
          type: string
        status:
          type: string
        closedAt:
          type: string
          format: date-time
        closedBy:
          type: string
        reopenedAt:
          type: string
          format: date-time
        reopenedBy:
          type: string
    AccountResponse:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        type:
          type: string
        currency:
          type: string
        parentCode:
          type: string
        active:
          type: boolean